refactor: extract ChEMBL tools into chembl.py
api_tools.py had grown to ~1660 lines, ~720 of them the ChEMBL cluster (3 tools
+ ~15 SPARQL/chemistry helpers). Move that block wholesale into a new
togo_mcp/chembl.py; main.py imports it (`from .chembl import *`) for the same
@mcp.tool registration side-effect.

Pure move, no logic changes. The split is clean because the earlier _rest_get
refactor already separated the shared REST plumbing: chembl.py imports the five
shared helpers it needs back from api_tools (_resolve_query_alias, _rest_get,
_RestError, _rest_fail_msg, _strip_html) and owns _chembl_client + everything
ChEMBL-specific. Nothing outside the block depended on the ChEMBL helpers.
Drops now-unused csv/io imports from api_tools.

ChEMBL tests likewise moved to tests/test_chembl.py (the backoff fixture patches
api_tools._REST_BACKOFF_BASE, where _rest_get reads it).

Result: api_tools.py 1660 → 940, new chembl.py 749. All 177 tests pass; all
three ChEMBL tools driven live end-to-end from the new module (aspirin→CHEMBL25,
EGFR→2 targets, imatinib→1).
